Fix #318
